Refactor: 페이지 분류
- 홈 페이지, 식당 조회 페이지, 투표 페이지로 분류
