+++ src/pages/HomePage/HomePage.css
@@ -1,0 +1,4 @@
+.home-page {
+  padding: 20px;
+  text-align: center;
+}
