+++ src/pages/HomePage/HomePage.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import './HomePage.css';
+
+const HomePage = () => {
+  return (
+    <div className="home-page">
+      <h2>뭐먹을까?</h2>
+      <p>뭐먹을까 메인 페이지입니다.</p>
+    </div>
+  );
+};
+
+export default HomePage;
