+++ src/pages/VotePage/VotePage.css
@@ -1,0 +1,4 @@
+.vote-page {
+  padding: 20px;
+  text-align: center;
+}
