+++ src/pages/VotePage/VotePage.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import './VotePage.css';
+
+const VotePage = () => {
+  return (
+    <div className="vote-page">
+      <h2>뭐먹을까?</h2>
+      <p>뭐먹을까 투표 페이지입니다.</p>
+    </div>
+  );
+};
+
+export default VotePage;
